Optimised BubbleImage class Removed MessagesGroupingCoordinator
